+++ src/Insider.Abstractions/IInsiderHookService.cs
@@ -1,0 +1,9 @@
+using System;
+using System.Reflection;
+
+namespace Insider;
+
+public interface IInsiderHookService
+{
+    IDisposable Detour(MethodInfo target, Delegate replacement);
+}